Not pretty but working.
Add a dependency on a meh library that does an okay job guessing gender. Since senator and rep names tend to be on the bland side, it actually works pretty well.

I tried to not use gender, but using 'them' in place of '[him/her]' sounded pretty bad. Also, filling in the name without a Mr. or Ms. didn't sound right either. So, I decided predicting gender made sense here. If we ever get a non-gender-binary representative, then it no longer works, but since we know that currently all names can accurately be classified as male or female, I think this is justified, in the name of making the filled-in scripts read better.

The scriptFormatter could probably be refactored so that the replacements are all part of one function, or a forEach loop, but it's late and I couldn't get that working.